+++ src/main/java/com/marcin/kontrahent_faktura_backend/controller/InvoiceController.java
@@ -32,22 +32,30 @@ public class InvoiceController {
 
     @RequestMapping(method = RequestMethod.GET, value = "/invoice/{id}")
     public InvoiceDto getInvoiceById(@RequestParam Long id) {
-        return mapper.mapToInvoiceDto(services.getInvoiceById(id));
+        if (services.checkIfInvoiceExists(id)) {
+            return mapper.mapToInvoiceDto(services.getInvoiceById(id));
+        } else throw new IllegalArgumentException("Invoice not found");
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/invoice", consumes = APPLICATION_JSON_VALUE)
     public InvoiceDto createInvoice(@RequestBody @Valid InvoiceDto invoiceDto) {
-        return mapper.mapToInvoiceDto(services.createInvoice(mapper.mapToInvoice(invoiceDto)));
+        if (!services.checkIfInvoiceExists(invoiceDto.getInvoiceId())) {
+            return mapper.mapToInvoiceDto(services.createInvoice(mapper.mapToInvoice(invoiceDto)));
+        } else throw new IllegalArgumentException("Invoice with this 'Id' already exists");
     }
 
     @RequestMapping(method = RequestMethod.PUT, value = "/invoice", consumes = APPLICATION_JSON_VALUE)
     public InvoiceDto updateInvoice(@RequestBody @Valid InvoiceDto invoiceDto) {
-        return mapper.mapToInvoiceDto(services.updateInvoice(mapper.mapToInvoice(invoiceDto)));
+        if (services.checkIfInvoiceExists(invoiceDto.getInvoiceId())) {
+            return mapper.mapToInvoiceDto(services.updateInvoice(mapper.mapToInvoice(invoiceDto)));
+        } else throw new IllegalArgumentException("Invoice not found");
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/invoice/{id}")
     public void deleteInvoice(@RequestParam Long id) {
-        services.deleteInvoice(id);
+        if (services.checkIfInvoiceExists(id)) {
+            services.deleteInvoice(id);
+        } else throw new IllegalArgumentException("Invoice not found");
     }
 
 }
